Start new command line interface for the Zephyr Scale API
This new command line interface (cli) uses argparse to parse
commands from the command line. It is now used to get test cases
only. But it is design to incorporate more functionality.

As part of the development, I updated the ``get_test_case`` method
to receive the ``raw`` argument. If it is false, some fields in the
payload will be parsed into more human readable format to easily
display in the love player tool